functions/slack_app/requirements.txt
Flask>1
google-cloud-firestore==2.5.3
slack-bolt==1.14.*
slack-sdk==3.18.*